+++ dsp/math/fast_math.c
@@ -43,7 +43,7 @@ int16_t q_tanh(int16_t x)
     int32_t a = ( (10 * x * x_squared) >> FIXED_POINT_Q ) + 105 * x;
     int32_t b = ( (x_squared * x_squared) >> FIXED_POINT_Q) + (45 * x_squared) + (105 << FIXED_POINT_Q);
 
-    return (a << FIXED_POINT_Q) / b;
+    return (int16_t) ((a << FIXED_POINT_Q) / b);
 }
 /*
 int main()
